database/requirements.txt
Flask 
Flask-Cors 
psycopg2-binary 
pyjwt 
bcrypt
cryptography

